Moved partner metadata classes to partner package
