+++ etor/src/main/java/gov/hhs/cdc/trustedintermediary/etor/orders/OrderConverter.java
@@ -1,7 +1,7 @@
 package gov.hhs.cdc.trustedintermediary.etor.orders;
 
 import gov.hhs.cdc.trustedintermediary.etor.demographics.Demographics;
-import gov.hhs.cdc.trustedintermediary.etor.metadata.PartnerMetadata;
+import gov.hhs.cdc.trustedintermediary.etor.metadata.partner.PartnerMetadata;
 import gov.hhs.cdc.trustedintermediary.etor.operationoutcomes.FhirMetadata;
 
 /** Interface for converting things to orders and things in orders. */
